pyproject.toml
[project]
name = "photodate"
version = "0.1.0"
description = "Corrigeer datums van ingescande foto's met AI"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.39.0",
    "typer>=0.9.0",
    "piexif>=1.1.3",
    "Pillow>=10.0.0",
    "rich>=13.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
]

[project.scripts]
photodate = "photodate.cli:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
